refactor: separate YLS demo
