+++ apps/demo/Drawer.tsx
@@ -1,21 +1,19 @@
 import { useState } from 'react';
 
+import { LogClick, LogScreen } from '@yourssu/logging-system-react';
 import { useLocation } from 'react-router-dom';
 
-import { LogClick } from '../components/LogClick';
-import { LogScreen } from '../components/LogScreen';
-
-export const Home = () => {
+export const Drawer = () => {
   const [count, setCount] = useState(0);
   const router = useLocation();
 
   return (
     <>
-      <h1>Home</h1>
+      <h1>Drawer</h1>
       <div className="card">
         <LogScreen
           params={{
-            userId: 'test',
+            userId: '',
             version: 1,
             event: {
               name: 'view',
@@ -25,11 +23,11 @@ export const Home = () => {
         >
           <LogClick
             params={{
-              userId: 'test',
-              version: 1,
+              userId: '',
+              version: 2,
               event: {
                 name: 'click',
-                screen: 'home',
+                screen: 'drawer',
               },
             }}
           >
