+++ apps/demo/Home.tsx
@@ -1,21 +1,19 @@
 import { useState } from 'react';
 
+import { LogClick, LogScreen } from '@yourssu/logging-system-react';
 import { useLocation } from 'react-router-dom';
 
-import { LogClick } from '../components/LogClick';
-import { LogScreen } from '../components/LogScreen';
-
-export const Drawer = () => {
+export const Home = () => {
   const [count, setCount] = useState(0);
   const router = useLocation();
 
   return (
     <>
-      <h1>Drawer</h1>
+      <h1>Home</h1>
       <div className="card">
         <LogScreen
           params={{
-            userId: '',
+            userId: 'test',
             version: 1,
             event: {
               name: 'view',
@@ -25,11 +23,11 @@ export const Drawer = () => {
         >
           <LogClick
             params={{
-              userId: '',
-              version: 2,
+              userId: 'test',
+              version: 1,
               event: {
                 name: 'click',
-                screen: 'drawer',
+                screen: 'home',
               },
             }}
           >
